+++ package.json
@@ -2,9 +2,9 @@
   "name": "@fastify/jwt",
   "version": "10.0.0",
   "description": "JWT utils for Fastify",
-  "main": "jwt.js",
+  "main": "index.js",
   "type": "commonjs",
-  "types": "types/jwt.d.ts",
+  "types": "types/index.d.ts",
   "scripts": {
     "lint": "eslint",
     "lint:fix": "eslint --fix",
