+++ test/jwt-async.test.js
@@ -2,7 +2,7 @@
 
 const { test } = require('node:test')
 const Fastify = require('fastify')
-const jwt = require('../jwt')
+const jwt = require('..')
 const { createSigner } = require('fast-jwt')
 
 test('Async key provider should be resolved internally', async function (t) {
